feat(foxy-query-builder): expose zoom value generator as a static method
